requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cryptography==43.0.1
cymem==2.0.8
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.1.0
fastapi==0.114.0
fastapi-cli==0.0.5
fastapi-sso==0.15.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pdfminer.six==20240706
pexpect==4.9.0
preshed==3.0.9
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
skillNer==1.0.3
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.4
thinc==8.2.5
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
websockets==13.0.1
wrapt==1.16.0
